Generate single JSON with all types of plans in the server (#1606)
* Moved plan simplification to the server

* Small fix in plan generation

* Minor fix in scripting

* Don't print OLAP plans yet

* Fixed unused arg problem

* fixed a typo

* Simplifying a copy of the plan now

* Fixed minor bugs

* fixed a bug in ut_common

* Cannonized some tests, changed plan comparison
