+++ ydb/tests/functional/canonical/canondata/test_sql.TestCanonicalFolder1.test_case_simple_script_params.script-script_/simple_script_params.script.plan
@@ -1,18 +1,21 @@
 {
     "meta": {
-        "version": "0.2",
-        "type": "script"
+        "type": "script",
+        "version": "0.2"
     },
     "queries": [
         {
-            "tables": [],
             "Plan": {
+                "Node Type": "Query",
+                "PlanNodeType": "Query",
                 "Plans": [
                     {
+                        "Node Type": "ResultSet",
                         "PlanNodeId": 2,
+                        "PlanNodeType": "ResultSet",
                         "Plans": [
                             {
-                                "PlanNodeId": 1,
+                                "Node Type": "ConstantExpr",
                                 "Operators": [
                                     {
                                         "Inputs": [],
@@ -20,16 +23,13 @@
                                         "Name": "Iterator"
                                     }
                                 ],
-                                "Node Type": "ConstantExpr"
+                                "PlanNodeId": 1
                             }
-                        ],
-                        "Node Type": "ResultSet",
-                        "PlanNodeType": "ResultSet"
+                        ]
                     }
-                ],
-                "Node Type": "Query",
-                "PlanNodeType": "Query"
-            }
+                ]
+            },
+            "tables": []
         }
     ]
 }
